% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS2dataPkg.R
\docType{data}
\name{hai}
\alias{hai}
\title{Descriptive data title}
\format{a \code{data.table} containing the following fields:
\describe{
\item{participant_id}{}
\item{age_reported}{}
\item{gender}{}
\item{race}{}
\item{cohort}{}
\item{study_time_collected}{}
\item{study_time_collected_unit}{}
\item{virus}{}
\item{value_reported}{}
\item{lsid}{}
}}
\source{
The data comes from ImmuneSpace via ImmuneSpaceR API.
}
\description{
Detailed description of the data
}
\seealso{
\link{IS2dataPkg}
\link{exprs}
}
