% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS2dataPkg.R
\docType{package}
\name{IS2dataPkg}
\alias{IS2dataPkg}
\alias{IS2dataPkg-package}
\title{IS2dataPkg}
\description{
Contains analysis-ready data for HIPC ImmuneSignatures2 Project
}
\details{
IS2dataPkg
A data package for IS2dataPkg.

Use \code{data(package='IS2dataPkg')$results[, 3]} to see a list of available data sets in this data package
    and/or DataPackageR::load_all_datasets() to load them.
}
\seealso{
\link{hai}
\link{exprs}
}
